feat : comment 로직 추가

+++ src/main/java/com/sparta/easyspring/post/PostService.java
@@ -1,5 +1,8 @@
 package com.sparta.easyspring.post;
 
+import com.sparta.easyspring.auth.entity.User;
+import com.sparta.easyspring.comment.dto.CommentResponseDto;
+import com.sparta.easyspring.comment.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -14,9 +17,10 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
+    private final CommentService commentService;
 
-    public PostResponseDto addPost(PostRequestDto requestDto) {
-        Post post = new Post(requestDto);
+    public PostResponseDto addPost(PostRequestDto requestDto, User user) {
+        Post post = new Post(requestDto,user);
         postRepository.save(post);
         return new PostResponseDto(post);
     }
@@ -31,18 +35,31 @@ public class PostService {
 
     public PostResponseDto getPost(Long postId) {
         Post post = findPostbyId(postId);
-        return new PostResponseDto(post);
+        PostResponseDto postResponseDto = new PostResponseDto(post);
+        List<CommentResponseDto> commentResponseDtoList = commentService.getAllComments(postId);
+        if(commentResponseDtoList.isEmpty()){
+            postResponseDto.setComments(null);
+        } else {
+            postResponseDto.setComments(commentResponseDtoList);
+        }
+        return postResponseDto;
     }
 
-    public PostResponseDto editPost(Long postId, PostRequestDto requestDto) {
+    public PostResponseDto editPost(Long postId, PostRequestDto requestDto, User user) {
         Post post = findPostbyId(postId);
+        if(!post.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("사용자 정보가 일치하지 않아 수정이 불가능합니다.");
+        }
         post.update(requestDto);
         postRepository.save(post);
         return new PostResponseDto(post);
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, User user) {
         Post post = findPostbyId(postId);
+        if(!post.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("사용자 정보가 일치하지 않아 삭제가 불가능합니다.");
+        }
         postRepository.delete(post);
     }
